// apps/frontend/src/types/database.ts
export interface Database {
  public: {
    Tables: {
// ...
      users: {
        Row: {
          id: string;
          email: string | null;
          avatar_url: string | null;
          status: string;
          first_name: string | null;
          last_name: string | null;
          primary_phone: string | null;
          secondary_phone: string | null;
          point_activated_at: string | null;
          created_at: string;
          updated_at: string;
        };
        Insert: {
          id?: string;
          email?: string | null;
          avatar_url?: string | null;
          status?: string;
          first_name?: string | null;
          last_name?: string | null;
          primary_phone?: string | null;
          secondary_phone?: string | null;
          point_activated_at?: string | null;
          created_at?: string;
          updated_at?: string;
        };
        Update: {
          id?: string;
          email?: string | null;
          avatar_url?: string | null;
          status?: string;
          first_name?: string | null;
          last_name?: string | null;
          primary_phone?: string | null;
          secondary_phone?: string | null;
          point_activated_at?: string | null;
          created_at?: string;
          updated_at?: string;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
}